+++ docs/CHANGELOG.md
@@ -1,1 +1,1 @@
-* Fix RS integration issue
+* Fix fluid duplication.